Finishes implementing checks for endorsements from authors
Issue: documentacao-e-tarefas/scielo#864

+++ classes/tasks/CheckEndorsements.php
@@ -3,10 +3,9 @@
 namespace APP\plugins\generic\plauditPreEndorsement\classes\tasks;
 
 use PKP\scheduledTask\ScheduledTask;
+use PKP\plugins\PluginRegistry;
 use APP\core\Application;
-use APP\submission\Submission;
 use Illuminate\Support\Facades\DB;
-use APP\plugins\generic\plauditPreEndorsement\classes\EndorsementService;
 use APP\plugins\generic\plauditPreEndorsement\classes\facades\Repo;
 use APP\plugins\generic\plauditPreEndorsement\classes\endorsement\Endorsement;
 
@@ -14,6 +13,8 @@ class CheckEndorsements extends ScheduledTask
 {
     public function executeActions()
     {
+        PluginRegistry::loadCategory('generic');
+        $plugin = PluginRegistry::getPlugin('generic', 'plauditpreendorsementplugin');
         $context = Application::get()->getRequest()->getContext();
         $endorsements = Repo::endorsement()->getCollector()
             ->filterByContextIds([$context->getId()])
@@ -25,7 +26,16 @@ class CheckEndorsements extends ScheduledTask
             $authorsEmails = $this->getPublicationAuthorsEmails($endorsement->getPublicationId());
 
             if (in_array($endorsement->getEmail(), $authorsEmails)) {
-                Repo::endorsement()->delete($endorsement);
+                $submissionId = $this->getSubmissionIdByEndorsement($endorsement);
+
+                if (!is_null($submissionId)) {
+                    Repo::endorsement()->delete($endorsement);
+                    $plugin->writeOnActivityLog(
+                        $submissionId,
+                        'plugins.generic.plauditPreEndorsement.log.endorsementRemoved.emailFromAuthor',
+                        ['email' => $endorsement->getEmail()]
+                    );
+                }
             }
         }
 
@@ -38,11 +48,22 @@ class CheckEndorsements extends ScheduledTask
             ->where('a.publication_id', $publicationId)
             ->select('a.email')
             ->get();
-        
+
         $emails = [];
         foreach ($rows as $row) {
             $emails[] = $row->email;
         }
         return $emails;
+    }
+
+    private function getSubmissionIdByEndorsement($endorsement)
+    {
+        $publicationId = $endorsement->getPublicationId();
+        $row = DB::table('publications AS p')
+            ->where('p.publication_id', $publicationId)
+            ->select('p.submission_id')
+            ->first();
+
+        return $row ? $row->submission_id : null;
     }
 }
